Move docs tasks to :documentation
