+++ documentation/documentation.gradle.kts
@@ -16,8 +16,14 @@ buildscript {
 
 plugins {
 	id("org.asciidoctor.convert")
+	id("org.ajoberstar.git-publish")
 	`kotlin-library-conventions`
 }
+
+val mavenizedProjects: List<Project> by rootProject.extra
+
+// Because we need to set up Javadoc aggregation
+mavenizedProjects.forEach { evaluationDependsOn(it.path) }
 
 javaLibrary {
 	mainJavaVersion = JavaVersion.VERSION_1_8
@@ -48,6 +54,30 @@ dependencies {
 
 asciidoctorj {
 	version = Versions.asciidoctorJ
+}
+
+val docsVersion = if (rootProject.version.toString().contains("SNAPSHOT")) "snapshot" else rootProject.version
+val docsDir = file("$buildDir/ghpages-docs")
+val replaceCurrentDocs = project.hasProperty("replaceCurrentDocs")
+val ota4jDocVersion = if (Versions.ota4j.contains("SNAPSHOT")) "snapshot" else Versions.ota4j
+val apiGuardianDocVersion = if (Versions.apiGuardian.contains("SNAPSHOT")) "snapshot" else Versions.apiGuardian
+
+gitPublish {
+	repoUri.set("https://github.com/junit-team/junit5.git")
+	branch.set("gh-pages")
+
+	contents {
+		from(docsDir)
+		into("docs")
+	}
+
+	preserve {
+		include("**/*")
+		exclude("docs/$docsVersion/**")
+		if (replaceCurrentDocs) {
+			exclude("docs/current/**")
+		}
+	}
 }
 
 val generatedAsciiDocPath = file("$buildDir/generated/asciidoc")
@@ -155,6 +185,106 @@ tasks {
 			}
 		}
 	}
+
+	val aggregateJavadocs by registering(Javadoc::class) {
+		dependsOn(mavenizedProjects.map { it.tasks.jar })
+		group = "Documentation"
+		description = "Generates aggregated Javadocs"
+
+		title = "JUnit $version API"
+
+		options {
+			memberLevel = JavadocMemberLevel.PROTECTED
+			header = rootProject.description
+			encoding = "UTF-8"
+			(this as StandardJavadocDocletOptions).apply {
+				splitIndex(true)
+				addBooleanOption("Xdoclint:none", true)
+				addBooleanOption("html5", true)
+				addBooleanOption("-no-module-directories", true)
+				addMultilineStringsOption("tag").value = listOf(
+						"apiNote:a:API Note:",
+						"implNote:a:Implementation Note:"
+				)
+				jFlags("-Xmx1g")
+				source("8") // https://github.com/junit-team/junit5/issues/1735
+				links("https://docs.oracle.com/javase/8/docs/api/")
+				links("https://ota4j-team.github.io/opentest4j/docs/$ota4jDocVersion/api/")
+				links("https://apiguardian-team.github.io/apiguardian/docs/$apiGuardianDocVersion/api/")
+				links("https://junit.org/junit4/javadoc/${Versions.junit4}/")
+				links("https://joel-costigliola.github.io/assertj/core-8/api/")
+				stylesheetFile = rootProject.file("src/javadoc/stylesheet.css")
+				groups = mapOf(
+						"Jupiter" to listOf("org.junit.jupiter.*"),
+						"Vintage" to listOf("org.junit.vintage.*"),
+						"Platform" to listOf("org.junit.platform.*")
+				)
+				use(true)
+				noTimestamp(true)
+			}
+		}
+		source(mavenizedProjects.map { it.sourceSets.main.get().allJava })
+
+		setMaxMemory("1024m")
+		setDestinationDir(file("$buildDir/docs/javadoc"))
+
+		classpath = files(mavenizedProjects.map { it.sourceSets.main.get().compileClasspath })
+				// Remove Kotlin classes from classpath due to "bad" class file
+				// see https://bugs.openjdk.java.net/browse/JDK-8187422
+				.filter { !it.path.contains("kotlin") }
+				// Remove subproject JARs so Kotlin classes don"t get picked up
+				.filter { it.isDirectory || !it.absolutePath.startsWith(projectDir.absolutePath) }
+
+		doLast {
+			// For compatibility with pre JDK 10 versions of the Javadoc tool
+			copy {
+				from(File(destinationDir, "element-list"))
+				into("$destinationDir")
+				rename { "package-list" }
+			}
+		}
+	}
+
+	val prepareDocsForUploadToGhPages by registering(Copy::class) {
+		dependsOn(aggregateJavadocs, asciidoctor)
+		outputs.dir(docsDir)
+
+		from("$buildDir/checksum") {
+			include("published-checksum.txt")
+		}
+		from("$buildDir/asciidoc") {
+			include("user-guide/**")
+			include("release-notes/**")
+			include("tocbot-*/**")
+		}
+		from("$buildDir/docs") {
+			include("javadoc/**")
+			filesMatching("**/*.html") {
+				val favicon = "<link rel=\"icon\" type=\"image/png\" href=\"https://junit.org/junit5/assets/img/junit5-logo.png\">"
+				filter { line ->
+					if (line.startsWith("<head>")) line.replace("<head>", "<head>$favicon") else line
+				}
+			}
+		}
+		into("$docsDir/$docsVersion")
+		filesMatching("javadoc/**") {
+			path = path.replace("javadoc/", "api/")
+		}
+		includeEmptyDirs = false
+	}
+
+	val createCurrentDocsFolder by registering(Copy::class) {
+		dependsOn(prepareDocsForUploadToGhPages)
+		outputs.dir("$docsDir/current")
+		onlyIf { replaceCurrentDocs }
+
+		from("$docsDir/$docsVersion")
+		into("$docsDir/current")
+	}
+
+	gitPublishCommit {
+		dependsOn(prepareDocsForUploadToGhPages, createCurrentDocsFolder)
+	}
 }
 
 fun redirectOutput(task: JavaExec, outputFile: File) {
@@ -168,9 +298,6 @@ fun redirectOutput(task: JavaExec, outputFile: File) {
 		}
 	}
 }
-
-evaluationDependsOn(":junit-platform-console")
-evaluationDependsOn(":junit-jupiter-params")
 
 eclipse {
 	classpath {
